refactor: migrate deck form page
