+++ src/features/deck/index.ts
@@ -1,2 +1,4 @@
-export { DeckFormContainer } from "./containers/DeckFormContainer";
+export { DeckForm, type DeckFormFields, type DeckFormProps } from "./components/DeckForm";
+export { useDeckActions } from "./hooks/useDeckActions";
 export { useDeckMutations } from "./hooks/useDeckMutations";
+export { deckFormSchema, type DeckFormValues } from "./lib/deckFormSchema";
