+++ src/pages/deck-form/ui/DeckFormView.stories.tsx
@@ -1,5 +1,5 @@
 /**
- * @file Defines Storybook examples for Deck Form Template.
+ * @file Defines Storybook examples for the Deck Form Page view.
  * These isolated scenarios show developers how the component looks, which props it accepts, and
  * how it responds to interaction.
  */
@@ -7,9 +7,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { INITIAL_VIEWPORTS } from "@/storybook/storybookViewports";
-import { DeckFormTemplate as Template } from "@/features/deck/components/templates/DeckFormTemplate";
-import type { DeckFormFields } from "@/features/deck/components/DeckForm";
+import type { DeckFormFields } from "@/features/deck";
 import * as fixture from "@/storybook/fixture";
+
+import { DeckFormView as Template } from "./DeckFormView";
 
 /**
  * Prepares fields for data for the Storybook examples in this file.
@@ -34,7 +35,7 @@ const longDeck: Deck = {
 };
 
 const meta = {
-  title: "Deck/DeckFormTemplate",
+  title: "Pages/Deck Form",
   component: Template,
   tags: ["autodocs"],
   parameters: {
